# CoW: Use weakref callbacks to track dead references