--- a/pandas/tests/copy_view/test_internals.py
+++ b/pandas/tests/copy_view/test_internals.py
@@ -115,31 +115,44 @@ def test_iset_splits_blocks_inplace(using_copy_on_write, locs, arr, dtype):
             assert not np.shares_memory(get_array(df, col), get_array(df2, col))
 
 
-def test_exponential_backoff():
-    # GH#55518
+def test_clear_dead_references():
+    # GH#55539
     df = DataFrame({"a": [1, 2, 3]})
     for i in range(490):
         df.copy(deep=False)
 
-    assert len(df._mgr.blocks[0].refs.referenced_blocks) == 491
+    assert (
+        len(df._mgr.blocks[0].refs.referenced_blocks)
+        - df._mgr.blocks[0].refs.dead_counter
+        == 1
+    )
 
     df = DataFrame({"a": [1, 2, 3]})
     dfs = [df.copy(deep=False) for i in range(510)]
 
     for i in range(20):
         df.copy(deep=False)
-    assert len(df._mgr.blocks[0].refs.referenced_blocks) == 531
-    assert df._mgr.blocks[0].refs.clear_counter == 1000
+    assert (
+        len(df._mgr.blocks[0].refs.referenced_blocks)
+        - df._mgr.blocks[0].refs.dead_counter
+        == 511
+    )
 
     for i in range(500):
         df.copy(deep=False)
 
-    # Don't reduce since we still have over 500 objects alive
-    assert df._mgr.blocks[0].refs.clear_counter == 1000
+    assert (
+        len(df._mgr.blocks[0].refs.referenced_blocks)
+        - df._mgr.blocks[0].refs.dead_counter
+        == 511
+    )
 
     dfs = dfs[:300]
     for i in range(500):
         df.copy(deep=False)
 
-    # Reduce since there are less than 500 objects alive
-    assert df._mgr.blocks[0].refs.clear_counter == 500
+    assert (
+        len(df._mgr.blocks[0].refs.referenced_blocks)
+        - df._mgr.blocks[0].refs.dead_counter
+        == 301
+    )