- [x] closes #55245
- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

This is an alternative solution to #55518. What it happening here:

- Uses the `weakref.ref` callback to count the number of dead references present at the `referenced_blocks` list.
- Takes 20% more time at the constructor because of the function declaration!
  - The function declaration can be moved to the `add_reference` and `add_index_reference`, but this is just moving the performance hit to these methods.
- Everything is O(1) now, including `has_reference`. (`_clear_dead_references` is amortized to O(1))
- If the number of dead references is more than 256 and +50% of the array, it will prune these references
   - The implication is that the linear slow dead reference cleaning happens next to the GC calls, which I think this is a more intuitive behavior. 

(also check another possibilities at #55008)

Let me know if I should continue working on this solution.