@WillAyd 2023-10-16
Assuming this is an alternate solution to @phofl PR in #55518