@phofl 2023-10-16
#55518 needs to be back ported. This is something that we can consider for main and 2.2, but I want to test this more before we rely on it.