@wangwillian0 2023-12-05
Sorry for the delay. Here are the benchmarks:

| Change   | Before [593fa855] <main>   | After [98addada] <weakref-callback>   |   Ratio | Benchmark (Parameter)                                          |
|----------|----------------------------|---------------------------------------|---------|----------------------------------------------------------------|
| +        | 5.74±0.2ms                 | 7.37±0.2ms                            |    1.29 | frame_ctor.FromArrays.time_frame_from_arrays_int               |
| +        | 7.65±0.08ms                | 8.63±0.2ms                            |    1.13 | frame_ctor.FromArrays.time_frame_from_arrays_sparse            |
| +        | 1.43±0.07ms                | 1.47±0.06ms                           |    1.03 | frame_ctor.FromRecords.time_frame_from_records_generator(1000) |